Remove the OpenAPI version check at importer instantiation time
The OpenAPI version is checked before the OpenAPIImporter is instantiated, so it is known to be valid. Additionally, the presence of quotations in the spec causes uploads to report successfully despite the absence of a mocked server. It looks like this is probably due to a file pointer overrun where isYaml remains null, though this has not been verified.